server/requirements.txt
# Install with: `pip install -r requirements.txt`

Django==1.11.3
psycopg2==2.7.1
django-extensions==1.8.1
gunicorn==19.7.1

# Admin
django-flat-responsive==1.4.1

# Security
django-admin-honeypot==1.0.0

# Rest Framework
djangorestframework==3.6.3
Markdown==2.6.8
django-filter==1.0.2

# CORS
django-cors-headers==2.1.0

# Logging
python-logstash==0.4.6

# Testing
locustio==0.8a2
